feat : Adding Migration & Updating Sqlmock

+++ simple-golang-database/internal/modules/product/model/entity/product.go
@@ -1,7 +1,7 @@
 package product_entity
 
 type Product struct {
-	ID    string  `db:"id"`
-	Name  string  `db:"name"`
-	Price float64 `db:"price"`
+	ID    string `db:"id"`
+	Name  string `db:"name"`
+	Price int64  `db:"price"`
 }
