+++ simple-golang-database/internal/modules/product/model/request/update_product.go
@@ -1,6 +1,6 @@
 package product_request
 
 type UpdateProduct struct {
-	Name  string  `json:"name" binding:"required"`
-	Price float64 `json:"price" binding:"required,gte=0"`
+	Name  string `json:"name" binding:"required"`
+	Price int64  `json:"price" binding:"required,gte=0"`
 }
